@@ -1,5 +1,5 @@
 {
-  "name": "myApp",
+  "name": "ottertherapies_ionic",
   "version": "0.0.1",
   "author": "Ionic Framework",
   "homepage": "https://ionicframework.com/",
@@ -9,17 +9,19 @@
     "build": "ng build",
     "test": "ng test",
     "lint": "ng lint",
-    "e2e": "ng e2e"
+    "e2e": "ng e2e",
+    "resources": "cordova-res ios && node scripts/resources.js"
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "~8.2.14",
-    "@angular/core": "~8.2.14",
+    "@angular/common": "~9.1.0",
+    "@angular/core": "~9.1.0",
     "@angular/fire": "^5.4.2",
-    "@angular/forms": "~8.2.14",
-    "@angular/platform-browser": "~8.2.14",
-    "@angular/platform-browser-dynamic": "~8.2.14",
-    "@angular/router": "~8.2.14",
+    "@angular/forms": "~9.1.0",
+    "@angular/platform-browser": "~9.1.0",
+    "@angular/platform-browser-dynamic": "~9.1.0",
+    "@angular/router": "~9.1.0",
+    "@angular/service-worker": "~9.1.0",
     "@capacitor/core": "^2.0.0-beta.1",
     "@capacitor/ios": "^2.0.0-beta.1",
     "@ionic-native/core": "^5.0.0",
@@ -31,7 +33,7 @@
     "@ionic-native/preview-any-file": "^5.22.0",
     "@ionic-native/splash-screen": "^5.0.0",
     "@ionic-native/status-bar": "^5.0.0",
-    "@ionic/angular": "^5.0.0",
+    "@ionic/angular": "^5.0.7",
     "cap": "^0.2.1",
     "com-sarriaroman-photoviewer": "^1.2.4",
     "cordova-plugin-document-scanner": "^4.2.3",
@@ -52,22 +54,23 @@
     "pdfmake": "^0.1.65",
     "rxjs": "^6.5.4",
     "sync": "^0.2.5",
-    "tslib": "^1.9.0",
+    "tslib": "^1.10.0",
     "uuid": "^7.0.1",
-    "zone.js": "~0.9.1"
+    "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.803.20",
-    "@angular/cli": "~8.3.23",
-    "@angular/compiler": "~8.2.14",
-    "@angular/compiler-cli": "~8.2.14",
-    "@angular/language-service": "~8.2.14",
+    "@angular-devkit/build-angular": "~0.901.0",
+    "@angular/cli": "~9.1.0",
+    "@angular/compiler": "~9.1.0",
+    "@angular/compiler-cli": "~9.1.0",
+    "@angular/language-service": "~9.1.0",
     "@capacitor/cli": "^2.0.0-beta.1",
     "@ionic/angular-toolkit": "^2.1.1",
     "@types/jasmine": "~3.3.8",
     "@types/jasminewd2": "~2.0.3",
-    "@types/node": "~8.9.4",
-    "codelyzer": "^5.0.0",
+    "@types/node": "^12.11.1",
+    "codelyzer": "^5.1.2",
+    "cordova-res": "^0.11.0",
     "jasmine-core": "~3.4.0",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "^4.4.1",
@@ -79,7 +82,7 @@
     "protractor": "~5.4.0",
     "ts-node": "~7.0.0",
     "tslint": "~5.15.0",
-    "typescript": "~3.4.3"
+    "typescript": "~3.8.3"
   },
   "description": "An Ionic project"
 }
