@@ -16,8 +16,8 @@ export const ROUTES: Routes = [
     path: 'auth',
     children: [
       { path: '', pathMatch: 'full', redirectTo: 'login' },
-      { path: 'login', loadChildren: './login/login.module#LoginModule' },
-      { path: 'register', loadChildren: './register/register.module#RegisterModule' },
+      { path: 'login', loadChildren: () => import('./login/login.module').then(m => m.LoginModule) },
+      { path: 'register', loadChildren: () => import('./register/register.module').then(m => m.RegisterModule) },
     ]
   }
 ];
